Fix Trigger Input Jack Issue
Previous version of these boards did not have the tip normal of the input jacks connected to +12v to allow free running mode. Kassutronics updated the documentation to add a resistor that prevents shorting in this connection. This update connects the tip normal in the input jacks to +12v through a 10k resistor.